requirements.txt
pandas
requests
python-dotenv
beautifulsoup4
lxml
scikit-learn
tabulate
tqdm
selenium
webdriver-manager
customtkinter